requirements.txt
Flask
Flask-Discord
stripe
python-dotenv
psycopg2-binary